+++ src/main/java/com/nhnacademy/minidooray/gateway/controller/common/AccountLoginController.java
@@ -3,11 +3,11 @@ package com.nhnacademy.minidooray.gateway.controller.common;
 import com.nhnacademy.minidooray.gateway.domain.account.request.AccountLoginRequestDTO;
 import com.nhnacademy.minidooray.gateway.domain.account.response.AccountStatusInfoResponseDTO;
 import com.nhnacademy.minidooray.gateway.service.account.AccountClientService;
+import com.nhnacademy.minidooray.gateway.service.account.impl.AccountClientServiceImpl;
 import java.util.Optional;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
-import javax.validation.ValidationException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
